requirements.txt
Flask
Celery
tensorflow
keras
spacy
nltk
pandas
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
msal
microsoftgraph-python
redis
pytest
coverage
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
